Pertemuan 7: Searching, Pagination, Filter, dan Show

+++ database/migrations/2026_04_08_071345_create_students_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('students', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('nim');
+            $table->bigInteger('nim');
             $table->timestamps();
         });
     }
